update(models): change type integer to bigInt

+++ db/models/journal.js
@@ -25,7 +25,7 @@ module.exports = (sequelize, DataTypes) => {
       description: DataTypes.STRING,
       id_coa: DataTypes.STRING,
       typeAmount: DataTypes.ENUM(['Debit', 'Kredit']),
-      amount: DataTypes.INTEGER,
+      amount: DataTypes.BIGINT,
     },
     {
       sequelize,
